requirements.txt
mcp-agent>=0.0.17
anthropic>=0.5.0
openai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0 